Resolve Issue #9
This commit solves the issue that arises when there is no contact with the bulbs. By implementing a kill function, we kill off the previous script if it's still hanging. If the current script needs to hang, it stores its PID in a file so the next script can kill it.