Finished refactoring controls.py to provide socket functionality
All of the functions in controls.py now use socket communication to communicate with the bulbs, bypassing the HTTP protocol used to communicate with the TP-link cloud.

All functions were tested using the debug controls and are working as intended.

Unable to test within a full script until initDev is changed to initialize devices.list to IP addresses.