Created bash script that saves local network connections to file
Created a bash script that pings the network, then uses arp to resolve the local network connections and saves them to a file. initDev() in controls.py can then parse this file for LB130 connections and save their IP addresses to file for use in the scripts. Then initDev will destroy the network list file