Fix for issue #20. (User override during transition)
This commit resolves #20. It implements a simple manual override detection which will prevent the script from changing the light, but still allow the script to update the default states of the bulb. This way, the next time the light is turned on, it will turn on to the appropriate setting.

This commit was tested and proven to work when both the brightness of the bulbs was changed during transition, and the hue. During both cases, the script continued uninterrupted, and the bulbs turned on to the appropriate setting the next time they were turned on.