Fix for issue #23 and relative file paths (#22)
This commit resolves #23 and resolves #22. The changes in the code allow the program to be much more portable.

I changed all the absolute file paths to relative file paths. I also made changes to getSun.py and getSunCron. The former of which will now also reside in the /etc/cron.d directory. getSun.py also locates the sunrise.py and sunset.py scripts prior to creating the cron file entries for those scripts.

This allows the user to change the installation location without affecting the program too much. (The scripts won't run until the next getSun runs).

As far as issue #23, this commit also fixes the issue with the bulbs not changing after shutoff at the switch for an extended period of time.

All code was tested and working in a controlled environment and will continue to be monitored over the next couple of days in the deployed environment.