Consolidation of transition scripts
In this commit, the transition scripts were refactored into one
transtion.py. This transition takes in two arguments. The first
argument is the state to transition to, and the second argument is the
interval to transition. Thus, I removed all of the previous transition
scripts (night.py, sunset.py, sunrise.py).

In addition, some utility functions were refactored out of transition.py
and added to their own file. This helped clean some things up.

I also modified the getSun.py script to reflect the usage of the new transition script.